Correct update in point.php
I'm very interested in using ECDH on a website, but the server where it will be hosted does not have GMP, only BCMath. I really liked your code and I'm trying to optimize it when using the bcmath. The replacement I made in the Mul method of the Point class managed save about 33% in some operations, how to generate public key or secret key. The original code was repeating the execution of bcmath_Utils::bcand.