Constant size for field element conditional swap
In the conditional swap algorithm, the values A and B
are swapped using a mask whose length is derived from
the bitsize of A or B (taking the largest)

This costs 12 `gmp_strval` AND `BinaryString::length`
operations per run of `Point::cswap`, so 24 per
iteration of the scalar multiplication loop.

The point values X/Y/order (bigints) are swapped,
along with isInfinity (a boolean). The mask for
isInfinity does not depend on the field size, however
X/Y/order have variable size.

This PR makes use of CurveParameters `size` value as
the mask size to avoid excessive string operations on
the values being swapped.

PointTest::testConditionalSwap yielded invalid results
with the change (without modifying the test). Its
unclear where the value of 32 for the size originated.
The test of cswap was a pure test, uncoupled with
the Point's values. It was updated to 67, the max bit
size of the two to get the test to pass.

Overall this change comes with some performance
improvements. With the point compression test from
`SpecBasedCurveTest::testGetPublicKey` separated out
so scalar multiplication was the dominant factor, the
test completed 14% faster with the change.

    Runs: 32
    A - using curve size
    AverageA: 8.5350606068969s
    TotalA: 273.1219394207s

    B - using existing method
    AverageB: 9.885170340538s
    TotalB: 316.32545089722s

    After 32 runs A is faster by 43.203511476517s
    A on average is faster by 1.3501097336411s seconds
    Time of A with respect to B: 86.342069108263%