Unit tests for random number generators
Adds a small test case checking that the classes
return something for the mcrypt and random_bytes
implementations.

The tests can technically fail, because decHex
won't pad the result16 bits, and there's a chance
that one byte would be 0x00, resulting in single
byte of output. It's unlikely, but we don't
generate that many test runs on 0.3